+++ buession-core/src/main/java/com/buession/core/validator/constraintvalidators/NotEmptyConstraintValidator.java
@@ -19,7 +19,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2022 Buession.com Inc.														       |
+ * | Copyright @ 2013-2026 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.core.validator.constraintvalidators;
@@ -27,8 +27,9 @@ package com.buession.core.validator.constraintvalidators;
 import com.buession.core.validator.Validate;
 import com.buession.core.validator.annotation.NotEmpty;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+
 import java.util.Collection;
 import java.util.Enumeration;
 import java.util.Iterator;
@@ -43,7 +44,7 @@ public abstract class NotEmptyConstraintValidator<T> implements ConstraintValida
 			extends NotEmptyConstraintValidator<CharSequence> {
 
 		@Override
-		public boolean isValid(CharSequence value, ConstraintValidatorContext context){
+		public boolean isValid(CharSequence value, ConstraintValidatorContext context) {
 			return Validate.isNotEmpty(value);
 		}
 
@@ -52,7 +53,7 @@ public abstract class NotEmptyConstraintValidator<T> implements ConstraintValida
 	public final static class ArrayNotEmptyConstraintValidator extends NotEmptyConstraintValidator<Object[]> {
 
 		@Override
-		public boolean isValid(Object[] value, ConstraintValidatorContext context){
+		public boolean isValid(Object[] value, ConstraintValidatorContext context) {
 			return Validate.isNotEmpty(value);
 		}
 
@@ -61,7 +62,7 @@ public abstract class NotEmptyConstraintValidator<T> implements ConstraintValida
 	public final static class MapNotEmptyConstraintValidator extends NotEmptyConstraintValidator<Map<?, ?>> {
 
 		@Override
-		public boolean isValid(Map<?, ?> value, ConstraintValidatorContext context){
+		public boolean isValid(Map<?, ?> value, ConstraintValidatorContext context) {
 			return Validate.isNotEmpty(value);
 		}
 
@@ -70,7 +71,7 @@ public abstract class NotEmptyConstraintValidator<T> implements ConstraintValida
 	public final static class CollectionNotEmptyConstraintValidator extends NotEmptyConstraintValidator<Collection<?>> {
 
 		@Override
-		public boolean isValid(Collection<?> value, ConstraintValidatorContext context){
+		public boolean isValid(Collection<?> value, ConstraintValidatorContext context) {
 			return Validate.isNotEmpty(value);
 		}
 
@@ -79,7 +80,7 @@ public abstract class NotEmptyConstraintValidator<T> implements ConstraintValida
 	public final static class IteratorNotEmptyConstraintValidator extends NotEmptyConstraintValidator<Iterator<?>> {
 
 		@Override
-		public boolean isValid(Iterator<?> value, ConstraintValidatorContext context){
+		public boolean isValid(Iterator<?> value, ConstraintValidatorContext context) {
 			return Validate.isNotEmpty(value);
 		}
 
@@ -89,7 +90,7 @@ public abstract class NotEmptyConstraintValidator<T> implements ConstraintValida
 			extends NotEmptyConstraintValidator<Enumeration<?>> {
 
 		@Override
-		public boolean isValid(Enumeration<?> value, ConstraintValidatorContext context){
+		public boolean isValid(Enumeration<?> value, ConstraintValidatorContext context) {
 			return Validate.isNotEmpty(value);
 		}