+++ buession-core/src/main/java/com/buession/core/validator/constraintvalidators/TelConstraintValidator.java
@@ -19,7 +19,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2022 Buession.com Inc.														       |
+ * | Copyright @ 2013-2026 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.core.validator.constraintvalidators;
@@ -28,8 +28,8 @@ import com.buession.core.validator.Validate;
 import com.buession.core.validator.annotation.Tel;
 import com.buession.core.validator.routines.TelValidator;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
 /**
  * @author Yong.Teng
@@ -39,12 +39,12 @@ public class TelConstraintValidator implements ConstraintValidator<Tel, CharSequ
 	protected TelValidator.AreaCodeType areaCodeType;
 
 	@Override
-	public void initialize(Tel tel){
+	public void initialize(Tel tel) {
 		this.areaCodeType = tel.areaCodeType();
 	}
 
 	@Override
-	public boolean isValid(CharSequence value, ConstraintValidatorContext context){
+	public boolean isValid(CharSequence value, ConstraintValidatorContext context) {
 		return Validate.isTel(value, areaCodeType);
 	}