+++ buession-core/src/main/java/com/buession/core/validator/constraintvalidators/XdigitConstraintValidator.java
@@ -19,7 +19,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2023 Buession.com Inc.														       |
+ * | Copyright @ 2013-2026 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.core.validator.constraintvalidators;
@@ -27,8 +27,8 @@ package com.buession.core.validator.constraintvalidators;
 import com.buession.core.validator.Validate;
 import com.buession.core.validator.annotation.Xdigit;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
 /**
  * @author Yong.Teng
@@ -38,7 +38,7 @@ public abstract class XdigitConstraintValidator<T> implements ConstraintValidato
 	public final static class CharSequenceXdigitConstraintValidator extends XdigitConstraintValidator<CharSequence> {
 
 		@Override
-		public boolean isValid(CharSequence value, ConstraintValidatorContext context){
+		public boolean isValid(CharSequence value, ConstraintValidatorContext context) {
 			return Validate.isXdigit(value);
 		}
 
@@ -47,7 +47,7 @@ public abstract class XdigitConstraintValidator<T> implements ConstraintValidato
 	public final static class CharXdigitConstraintValidator extends XdigitConstraintValidator<Character> {
 
 		@Override
-		public boolean isValid(Character value, ConstraintValidatorContext context){
+		public boolean isValid(Character value, ConstraintValidatorContext context) {
 			return Validate.isXdigit(value);
 		}